refactor: add sdk runtime
Similar to react-sdk/runtime added another entry point to sdk package
which includes all the stuff imported on published sites to reduce
js size.
