+++ packages/sdk/src/index.ts
@@ -14,10 +14,7 @@ export * from "./schema/webstudio";
 export * from "./instances-utils";
 export * from "./page-utils";
 export * from "./scope";
-export * from "./resource-loader";
 export * from "./expression";
 export * from "./resources-generator";
 export * from "./page-meta-generator";
 export * from "./url-pattern";
-export * from "./form-fields";
-export * from "./to-string";
