+++ packages/sdk/src/resource-loader.ts
@@ -1,5 +1,22 @@
 import type { ResourceRequest } from "./schema/resources";
 
+const LOCAL_RESOURCE_PREFIX = "$resources";
+
+/**
+ * Prevents fetch cycles by prefixing local resources.
+ */
+export const isLocalResource = (pathname: string, resourceName?: string) => {
+  const segments = pathname.split("/").filter(Boolean);
+
+  if (resourceName === undefined) {
+    return segments[0] === LOCAL_RESOURCE_PREFIX;
+  }
+
+  return segments.join("/") === `${LOCAL_RESOURCE_PREFIX}/${resourceName}`;
+};
+
+export const sitemapResourceUrl = `/${LOCAL_RESOURCE_PREFIX}/sitemap.xml`;
+
 export const loadResource = async (
   customFetch: typeof fetch,
   resourceRequest: ResourceRequest
