+++ packages/sdk/src/schema/resources.ts
@@ -44,20 +44,3 @@ export type ResourceRequest = z.infer<typeof ResourceRequest>;
 export const Resources = z.map(ResourceId, Resource);
 
 export type Resources = z.infer<typeof Resources>;
-
-const LOCAL_RESOURCE_PREFIX = "$resources";
-
-/**
- * Prevents fetch cycles by prefixing local resources.
- */
-export const isLocalResource = (pathname: string, resourceName?: string) => {
-  const segments = pathname.split("/").filter(Boolean);
-
-  if (resourceName === undefined) {
-    return segments[0] === LOCAL_RESOURCE_PREFIX;
-  }
-
-  return segments.join("/") === `${LOCAL_RESOURCE_PREFIX}/${resourceName}`;
-};
-
-export const sitemapResourceUrl = `/${LOCAL_RESOURCE_PREFIX}/sitemap.xml`;
